macro encapsulation
