+++ rust/common/mod.rs
@@ -26,4 +26,6 @@ pub mod string;
 pub mod token;
 pub mod validatable;
 
-pub type Result<T> = std::result::Result<T, error::Error>;
+pub use error::Error;
+
+pub type Result<T> = std::result::Result<T, Error>;
